+++ R/window.piar_index.R
@@ -1,0 +1,57 @@
+#' Get the indexes for a window of time periods
+#' @noRd
+index_window <- function(x, start, end) {
+  start <- match(
+    if (is.null(start)) start(x) else match.arg(start, x$time),
+    x$time
+  )
+  end <- match(
+    if (is.null(end)) end(x) else match.arg(end, x$time),
+    x$time
+  )
+  
+  if (start > end) {
+    stop("'start' must refer to a time period before 'end'")
+  }
+  seq.int(start, end)
+}
+
+#' Index window
+#' 
+#' Extract and replace index values over a window of time periods.
+#' 
+#' @param x A price index, as made by, e.g., [elemental_index()].
+#' @param start The time period to start the window. The default in the first
+#' period of `x`.
+#' @param end The time period to end the window. The default is the last period
+#' of `x`.
+#' @param ... Not currently used.
+#' @param value A numeric vector or price index.
+#' 
+#' @returns
+#' `window()` extracts a price index over a window of time periods that
+#' inherits from the same class as `x`. The replacement method replaces these
+#' with `value`.
+#' 
+#' @examples
+#' x <- as_index(matrix(1:9, 3))
+#' 
+#' window(x, "2")
+#' 
+#' window(x, "2") <- 1
+#' x
+#' 
+#' @family index methods
+#' @importFrom stats window
+#' @export
+window.piar_index <- function(x, start = NULL, end = NULL, ...) {
+  x[, index_window(x, start, end)]
+}
+
+#' @rdname window.piar_index
+#' @importFrom stats window<-
+#' @export
+`window<-.piar_index` <- function(x, start = NULL, end = NULL, ..., value) {
+  x[, index_window(x, start, end)] <- value
+  x
+}